Refactoring and renaming of 3D consistency classes and wedgefilter classes into a combined fourierConsistency package
