+++ src/edu/stanford/rsl/tutorial/fourierConsistency/wedgefilter/DoubleWedgeFilterFanEA.java
@@ -2,7 +2,7 @@
  * Copyright (C) 2014 Marcel Pohlmann
  * CONRAD is developed as an Open Source project under the GNU General Public License (GPL).
 */
-package edu.stanford.rsl.tutorial.wedgefilter;
+package edu.stanford.rsl.tutorial.fourierConsistency.wedgefilter;
 
 import java.util.ArrayList;
